+++ test/models/licence.model.test.js
@@ -12,12 +12,16 @@ const BillLicenceHelper = require('../support/helpers/bill-licence.helper.js')
 const BillLicenceModel = require('../../app/models/bill-licence.model.js')
 const ChargeVersionHelper = require('../support/helpers/charge-version.helper.js')
 const ChargeVersionModel = require('../../app/models/charge-version.model.js')
+const CompanyHelper = require('../support/helpers/company.helper.js')
+const ContactHelper = require('../support/helpers/contact.helper.js')
 const DatabaseHelper = require('../support/helpers/database.helper.js')
 const LicenceHelper = require('../support/helpers/licence.helper.js')
 const LicenceDocumentHelper = require('../support/helpers/licence-document.helper.js')
 const LicenceDocumentModel = require('../../app/models/licence-document.model.js')
 const LicenceDocumentHeaderHelper = require('../support/helpers/licence-document-header.helper.js')
 const LicenceDocumentHeaderModel = require('../../app/models/licence-document-header.model.js')
+const LicenceDocumentRoleHelper = require('../support/helpers/licence-document-role.helper.js')
+const LicenceRoleHelper = require('../support/helpers/licence-role.helper.js')
 const LicenceVersionHelper = require('../support/helpers/licence-version.helper.js')
 const LicenceVersionModel = require('../../app/models/licence-version.model.js')
 const RegionHelper = require('../support/helpers/region.helper.js')
@@ -458,6 +462,102 @@ describe('Licence model', () => {
           const result = LicenceModel.fromJson({ expiredDate, lapsedDate, revokedDate }).$ends()
 
           expect(result).to.equal({ date: new Date('2023-03-09'), priority: 1, reason: 'revoked' })
+        })
+      })
+    })
+  })
+
+  describe('$licenceHolder', () => {
+    describe('when instance has not been set with the additional properties needed', () => {
+      it('returns null', () => {
+        const result = testRecord.$licenceHolder()
+
+        expect(result).to.be.null()
+      })
+    })
+
+    describe('when the instance has been set with the additional properties needed', () => {
+      const licenceRoles = {}
+
+      let licence
+      let company
+      let contact
+      let licenceDocument
+
+      beforeEach(async () => {
+        licence = await LicenceHelper.add()
+
+        // Create 2 licence roles so we can test the service only gets the licence document role record that is for
+        // 'licence holder'
+        licenceRoles.billing = await LicenceRoleHelper.add({ name: 'billing', label: 'Billing' })
+        licenceRoles.holder = await LicenceRoleHelper.add()
+
+        // Create company and contact records. We create an additional company so we can create 2 licence document role
+        // records for our licence to test the one with the latest start date is used.
+        company = await CompanyHelper.add({ name: 'Licence Holder Ltd' })
+        contact = await ContactHelper.add({ firstName: 'Luce', lastName: 'Holder' })
+        const oldCompany = await CompanyHelper.add({ name: 'Old Licence Holder Ltd' })
+
+        // We have to create a licence document to link our licence record to (eventually!) the company or contact
+        // record that is the 'licence holder'
+        licenceDocument = await LicenceDocumentHelper.add({ licenceRef: licence.licenceRef })
+
+        // Create two licence document role records. This one is linked to the billing role so should be ignored by the
+        // service
+        await LicenceDocumentRoleHelper.add({
+          licenceDocumentId: licenceDocument.id,
+          licenceRoleId: licenceRoles.billing.id
+        })
+
+        // This one is linked to the old company record so should not be used to provide the licence holder name
+        await LicenceDocumentRoleHelper.add({
+          licenceDocumentId: licenceDocument.id,
+          licenceRoleId: licenceRoles.holder.id,
+          company: oldCompany.id,
+          startDate: new Date('2022-01-01')
+        })
+      })
+
+      describe('and the licence holder is a company', () => {
+        beforeEach(async () => {
+          // Create the licence document role record that _is_ linked to the correct licence holder record
+          await LicenceDocumentRoleHelper.add({
+            licenceDocumentId: licenceDocument.id,
+            licenceRoleId: licenceRoles.holder.id,
+            companyId: company.id,
+            startDate: new Date('2022-08-01')
+          })
+
+          testRecord = await LicenceModel.query().findById(licence.id).modify('licenceHolder')
+        })
+
+        it('returns the company name as the licence holder', async () => {
+          const result = testRecord.$licenceHolder()
+
+          expect(result).to.equal('Licence Holder Ltd')
+        })
+      })
+
+      describe('and the licence holder is a contact', () => {
+        beforeEach(async () => {
+          // Create the licence document role record that _is_ linked to the correct licence holder record.
+          // NOTE: We create this against both the company and contact to also confirm that the contact name has
+          // precedence over the company name
+          await LicenceDocumentRoleHelper.add({
+            licenceDocumentId: licenceDocument.id,
+            licenceRoleId: licenceRoles.holder.id,
+            companyId: company.id,
+            contactId: contact.id,
+            startDate: new Date('2022-08-01')
+          })
+
+          testRecord = await LicenceModel.query().findById(licence.id).modify('licenceHolder')
+        })
+
+        it('returns the contact name as the licence holder', async () => {
+          const result = testRecord.$licenceHolder()
+
+          expect(result).to.equal('Luce Holder')
         })
       })
     })
